feat(html): add view trace button to top of test view
