+++ packages/html-reporter/src/testCaseView.css
@@ -39,7 +39,6 @@
   flex: none;
   align-items: center;
   padding: 0 8px 8px;
-  line-height: 24px;
 }
 
 .test-case-run-duration {
