+++ packages/html-reporter/src/testFileView.css
@@ -51,28 +51,6 @@
   margin-right: 10px;
 }
 
-.test-file-badge {
-  flex: none;
-  background-color: transparent;
-  border-color: transparent;
-}
-
-.test-file-badge span {
-  color: var(--color-fg-muted);
-}
-
-.test-file-badge:hover {
-  cursor: pointer;
-}
-
-.test-file-badge svg {
-  fill: var(--color-fg-muted);
-}
-
-.test-file-badge:hover svg {
-  fill: var(--color-fg-muted);
-}
-
 .test-file-test-outcome-skipped {
   color: var(--color-fg-muted);
 }
